Move relevant tests to Activout.RestClient.Newtonsoft.Json.Test
