+++ Activout.RestClient.Newtonsoft.Json.Test/DomainExceptions/MyDomainHttpErrorAttribute.cs
@@ -1,8 +1,9 @@
 using System.Net;
+using Activout.RestClient.DomainExceptions;
 
-namespace Activout.RestClient.Test.DomainExceptionTests
+namespace Activout.RestClient.Newtonsoft.Json.Test.DomainExceptions
 {
-    internal class MyDomainHttpErrorAttribute : DomainExceptions.DomainHttpErrorAttribute
+    internal class MyDomainHttpErrorAttribute : DomainHttpErrorAttribute
     {
         public MyDomainHttpErrorAttribute(HttpStatusCode httpStatusCode, MyDomainErrorEnum domainErrorValue) : base(
             httpStatusCode, domainErrorValue)
