+++ Activout.RestClient.Test/MultipartFormDataContentTest.cs
@@ -3,7 +3,6 @@ using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
-using Activout.RestClient.Newtonsoft.Json;
 using Microsoft.Extensions.Logging;
 using RichardSzalay.MockHttp;
 using Xunit;
