+++ ActivoutRestClient.sln
@@ -13,6 +13,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Activout.RestClient.Xml.Tes
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Activout.RestClient.Newtonsoft.Json", "Activout.RestClient.Newtonsoft.Json\Activout.RestClient.Newtonsoft.Json.csproj", "{0D33957E-E4C6-4C17-AEBA-6D6B2A90F1CB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Activout.RestClient.Newtonsoft.Json.Test", "Activout.RestClient.Newtonsoft.Json.Test\Activout.RestClient.Newtonsoft.Json.Test.csproj", "{D4DE7929-0435-46AB-93F3-5615536CA08D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -39,6 +41,10 @@ Global
 		{0D33957E-E4C6-4C17-AEBA-6D6B2A90F1CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0D33957E-E4C6-4C17-AEBA-6D6B2A90F1CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0D33957E-E4C6-4C17-AEBA-6D6B2A90F1CB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D4DE7929-0435-46AB-93F3-5615536CA08D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D4DE7929-0435-46AB-93F3-5615536CA08D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D4DE7929-0435-46AB-93F3-5615536CA08D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D4DE7929-0435-46AB-93F3-5615536CA08D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
